Implemented Auth using Policies

+++ app/Http/Controllers/QuestionController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionController extends Controller
 {
+    public function __construct() {
+      $this->middleware('auth',['except' => ['index','show']]);
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -68,9 +72,13 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        /*
         if(\Gate::denies('update-question',$question)){
             abort(403,"Access denied");
         }
+        */
+
+        $this->authorize("update",$question);
 
         return view("questions.edit",compact('question'));
 
@@ -85,9 +93,13 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        /*
         if(\Gate::denies('update-question',$question)){
             abort(403,"Access denied");
         }
+        */
+
+        $this->authorize("update",$question);
 
         $question->update($request->only('title','body'));
 
@@ -102,9 +114,13 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        /*
         if(\Gate::denies('delete-question',$question)){
             abort(403,"Access denied");
         }
+        */
+
+        $this->authorize("delete",$question);
 
         $question->delete();
